requirements.txt
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
wtforms-validators==1.0.0
zipp==3.17.0
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
wtforms-validators==1.0.0
zipp==3.17.0
